requirements.txt
streamlit
pandas
plotly
numpy
python-dotenv
requests
streamlit-option-menu